Biggest commit yet, plugin is almost done:
Created an event that runs a function in SpleefGame every tick when a game is active, checking if players have fallen out and if the game is over.

Made SerializableInventory store armor data, as well as change where it was referenced elsewhere to match.

Made it so it gives every player a shovel when the countdown ends.